Enable vertical coarsening of additional variables in data processing (#1649)
Our data processing script is currently fairly rigid about the variables
it vertically coarsens, doing so only for fields hard-coded in the
`vertically_resolved` attribute of the `StandardNameMapping` class:
https://github.com/ai2cm/full-model/blob/15ea6c4fdf78ad92fe038f95f541e450f3b50461/scripts/data_process/compute_dataset.py#L96-L101

We have started to generate datasets with additional 3D fields, which
require vertical coarsening. This PR adds an
`additional_vertically_resolved_names` attribute to the
`StandardNameMapping` class, which allows a user to specify names of
additional variables present in their dataset that they would like to
see vertically coarsened.

Changes:
- Adds an `additional_vertically_resolved_names` attribute to the
`StandardNameMapping` class in `compute_dataset`.
- Adds a configuration for processing the latest increasing CO2 dataset,
which is an example of using this new functionality.