Bump `n_split` up in increasing CO2 data processing (#1651)
This looks like it is going to be quite slow—it is on pace to finish in
over 3 days—but it will finish. For these kinds of long runs it would be
useful to be able to split across multiple workers in an array-type job,
since we don't have the ability to parallelize over ensemble member.

This PR also fixes a typo in the `Makefile`.

Argo job: `compute-fme-dataset-ensemble-7lvdt`.