Reindex annual-mean data to all years for clearer plots (#1655)
Currently our annual- and global-mean plots are confusing because the
ensemble mean line connects between years that correspond to entirely
different runs. This is because the 'year' coordinate is often something
like [2010, 2011, 2012, 2020, 2021, 2022] etc. 

This updates the annual aggregator plotting code to re-index to a series
of all years between the 'min' and 'max' year, which will automatically
insert NaNs for years for which there is no data. I verified in a
notebook that this leads to the intended effect in the plots, although
have not actually verified this in a production run.