Generalize `AtmosphereData.frozen_precipitation_rate` (#1652)
Previously `AtmosphereData.frozen_precipitation_rate` required the
presence of three SHiELD-specific diagnostics (`"GRAUPELsfc"`,
`"ICEsfc"`, and `"SNOWsfc"`) and returned their sum. We are moving
towards predicting a single total frozen precipitation rate for
simplicity and generalization to other models / datasets.

Changes:
- Updates `AtmosphereData.frozen_precipitation_rate` to first look for a
single key representing the total frozen precipitation rate, then fall
back to returning the sum of the components if present, and finally fall
back to returning a tensor of zeros if it is not possible to determine
the frozen precipitation rate given the data.

- [x] Tests added