Include beaker-py in `atmosphere-processing` image (#1656)
Now that we have decoupled the image we use for atmosphere data
processing from fv3net, it is easier to add dependencies. Therefore,
instead of maintaining a separate small image for beaker-py, we can
include it in the `atmosphere-processing` image and use that for all
steps of the workflow. This resolves the issue that the beaker-py image
is currently hosted on Container Registry, which is being deprecated in
about a week.

The `upload-beaker-stats` step ran successfully in this small test job:
`compute-fme-dataset-ensemble-bfh7f`.

Changes:
- Installs beaker-py into the `atmosphere-processing` image and uses
that in the `upload-beaker-stats` workflow step.
 
If we wanted we could merge the `upload-beaker-stats` step into the
`combine-stats` step, since that can all be done in the same image now.