Refactor SingleModuleStepper to depend on StepABC (#1628)
This PR refactors the SingleModuleStepper so it takes in a StepABC as
input argument, instead of a SingleModuleStepper.

In practice, no behavior is changed, and a SingleModuleStepper is still
always passed as this argument. However this will make it much easier to
change this behavior in another PR without having to combine it with the
large number of refactor changes needed in this one.

Changes:
- Removed `SingleModuleStepper.timestep`, replacing it with
`validate_inference_data`
- Removed unused `diagnostic_names` and `in_names` attributes from
SingleModuleStepper
- Replaced some name attributes with manual name arithmetic inside
`_predict_generator`, where the logic is clearer. The context prevents
ambiguity of "forcing" meaning either input-only variables provided on
the input step or forcing variables provided on the output step.
- Added `atmosphere_timestep` and `ocean_timestep` attributes to coupled
TrainBuilders, to avoid trying to get this information from the built
atmosphere/ocean steppers.

- [ ] Tests added