Combine loss normalization with loss object (#1646)
This PR updates the loss object we use to take in denormalized data.

This will make it easier to dependency-inject the loss object to init
instead of building it, as computing the loss will involve only one
external object instead of two. Some other changes need to happen before
we can do this (because of coupling of the loss with multi-call).

Changes:
- `MappingLoss` renamed to `WeightedMappingLoss`, which now takes in
denormalized data

- [ ] Tests added