Improve performance of atmosphere data processing and stats (#1671)
This PR makes some updates to improve the performance of our data
processing and stats workflows. The main changes are switching to use
nodes in the `highmem-sim-pool` for these steps rather than the
`med-sim-pool`, and using the distributed scheduler instead of the
single-machine scheduler for the computation.

Changes:
- Use the `highmem-sim-pool` for the `compute-fme-dataset-individual`
and `get-stats` steps of the data processing workflow.
- Use the distributed scheduler with a hard-coded 16 workers instead of
the single-machine scheduler in the `compute-fme-dataset-individual` and
`get-stats` steps.
- Add manual timing print outs to track timing and progress, since
dask's `ProgressBar`, which we used previously, only works with the
single-machine scheduler. In practice, due to the way logging works in
the cloud, we would never see the progress bar until after the segment
was complete anyway.
- Switches from `print` statements to `logging.info` statements, since
the former would only be displayed after all segments were complete when
using the distributed scheduler.

Empirically making these changes provides a roughly 4x speed up for
computing the dataset and the stats:

| Step    | Before | After | Speed up |
| -------- | ------- |----|----|
| `compute-fme-dataset-individual` seconds per shard (3 months) | 884 |
208 | 4.25x |
|`get-stats` for one year of data |  1700  | 418 | 4.07x |

Given that the n2-highmem-32 nodes of our `highmem-sim-pool` are about
4x as expensive per hour as the e2-standard-16 nodes of our
`med-sim-pool` ([pricing
link](https://cloud.google.com/compute/all-pricing?hl=en)), this should
hopefully mean our compute costs do not increase, but our time to result
decreases.

We could consider adding bokeh to our image, which is needed for viewing
the dask distributed dashboard. I installed it manually via `pip` while
investigating performance, and viewed it by connecting to the running
kubernetes pod, e.g.:

```
$ kubectl port-forward compute-fme-dataset-ensemble-zg48r-402608688 8787
```

For now though, I think it is OK for this to be done manually on an as
needed basis.