Consolidate dataset information into provider object (#1665)
This PR consolidates dataset information (gridded operations,
coordinates, timestep, img_shape) into a provider object which provides
these attributes if available, or gives an error informing the user the
dataset is missing required information.

The PR makes use of it within the new StepABC system. The existing
SingleModuleStepper system is compatible with it, as it has the
information needed to construct a DatasetInfo object.

Changes:
- Added DatasetInfo class containing training dataset information used
by Step objects
- Replaced img_shape, gridded_operations, vertical_coordinate and
timestep arguments to Step-related routines with dataset_info
- Implemented `__eq__` for GriddedOperations

- [x] Tests added