Re-process new increasing CO2 dataset (#1687)
Due to a filesystem issue on Gaea, the new increasing CO2 simulation had
actually stopped running with three years to go, so our previously
processed version was not complete. We restarted the run, ran the
remaining three years, and re-processed the output. This PR updates our
config with the new dataset paths.

Changes:
- Updates the
`shield-som-increasing-co2-with-3d-heating-rate-diagnostics-1deg-8layer.yaml`
config with new paths for the raw and processed versions of the complete
new increasing CO2 run.

The processing job, `compute-fme-dataset-ensemble-mdqgn`, finished
successfully over the weekend. I deleted the previous incomplete raw and
processed datasets associated with this run on GCS.