Move NaN-filling to after time-subsetting in data loader (#1686)
This PR addresses #1685 by moving NaN-filling to after time-subsetting
within the data loader. I submitted a test production job
[here](https://beaker.allen.ai/orgs/ai2/workspaces/ace/work/01JPJ6R82VPYDYS7KY9SFXVR55/logs?jobId=01JPJ6R87811PKQK8QH8RJMZXQ),
which seems to have made it through an entire training epoch
successfully. The configuration is identical to the failed job
referenced in #1685.

Changes:
- Makes `fill_nans` an optional argument in `load_series_data` and
performs NaN-filling after time-subsetting and loading the data.

- [x] Tests added (a bit tricky to explicitly test the underlying issue,
since we expect the final data to be loaded into memory whether we fill
NaNs before or after time-subsetting, but the existing NaN-filling tests
pass).

Resolves #1685