Fail faster on invalid module config (#1693)
Currently if an invalid config is provided to `ModuleSelector` our
training code does not fail until the initialization of the stepper.
This PR adds a call to ensure that the concrete module config can be
built from the provided `type: str` and `config: Mapping[str, Any]`
during post_init of the module selector. This will ensure that the
`fme.ace.validate_config` entrypoint will catch these types of issues.

Resolves #1692