Add SingleModuleStep and SingleModuleStepConfig complying with ABCs (#1690)
This PR adds SingleModuleStep and SingleModuleStepConfig to be used with
the new step ABCs.

This will be regression tested against the existing single module code
in a later PR, when it is hooked into the train and inference system
with backwards compatibility conversions from the old to new-style
config.

Changes:
- Added `SingleModuleStep` and `SingleModuleStepConfig` in `fme.core`

- [x] Tests added