Move parameter init to stepper (#1712)
This PR moves ParameterInitializationConfig to the Stepper, instead of
on each Step object.

The API of ParameterInitializationConfig is general and can be applied
to any `.modules`. This change prevents us from having to duplicate the
config on each step object.

Changes:
- ParameterInitializationConfig no longer returns the modules it is
modifying in-place, making it clear that it is doing in-place operations
- The leading `module` is no longer stripped when loading weights to
apply or when applying weights to the destination, simplifying the logic
- `init_weights` is now an argument on the Stepper init instead of the
Step init, as it determines whether ParameterInitializationConfig is
applied
- Removed unused `get_base_weights` routines from step objects

- [ ] Tests added