Move stepper checkpoint handling to stepper module (#1711)
Currently the coupled stepper.py does low-level loading of vertical
coordinates from an ace stepper checkpoint. This PR moves the low-level
operations to the ace stepper modules, instead using a public interface.
This will make it easier to update the format of the checkpoint in a
later PR.

Changes:
- Add `get_serialized_stepper_vertical_coord` and use it from the
coupled stepper.py

- [x] Tests added