Make loss initialization lazy in stepper (#1714)
A later PR will ensure that the stepper cannot go out-of-sync from its
config by using reinitialization from config as the main way of
modifying or reloading the object. Because of
https://github.com/ai2cm/full-model/issues/1698, this would cause the
stepper to crash whenever using multi-call due to statistics being
missing, even though we don't actually need to use the loss.

This PR updates loss initialization to be lazy, so that when replacing a
stepper's multi-call config we will not get an error unless the loss is
actually used.

Changes:
- Updated loss initialization in Stepper to be lazy
- Moved loss initialization logic to SingleModuleStepperConfig

- [ ] Tests added