Vertically coarsen land variables (#1679)
LM4 has 20 levels of total moisture in a layer and soil temperature, for
the land model emulator we want to vertically coarsen this to match
ERA5's 4 levels. For the soil temperature, using height thickness-based
weight mean for the coarsening is appropriate and for total moisture of
a soil level using a simple sum over the layers will conserve volumetric
water content. This PR changes `compute_dataset.py` to allow this.

Changes:
- Adds height-based thickness calculation for the vertical coordinates
of the land model in `compute_dataset.py`
- Allows for the passing of a separate list of indices to coarsen over
for the 3D land variables in `compute_dataset.py`
- Adds function to do the vertical coarsening of 3D land variables
- Changes existing config (CM4-piControl-atmosphere-1deg-8layer-200yr.yaml) to allow for CM4 atmosphere + 3D land variables

---------

Co-authored-by: Spencer Clark <spencerkclark@gmail.com>