Non-static area weights for downscaling (#1722)
This PR is a component of the downscaling rewrite that removes the
assumption of a static spatial locations.

The area weights are another field that propagated largely through the
aggregators and models under the assumption that they were static across
the entire training/evaluation loop. This PR removes that assumption by
largely removing that information from the models and concentrating the
ability to dynamically assign a metric (e.g., loss) a set of area
weights for the aggregation.

Changes:
- Defined a new aggregator protocal
`_DynamicMetricComparisonAggregator`, which adds an optional
`dynamic_metric` argument for a callable to define a per-batch metric.
- `MeanComparison` uses this new protocol, and I use it to override an
area-weighted metric in the `Aggregator.record_batch` method
- `Aggregator.record_batch` now requires an `area_weights` tensor
- Adjusted `LossConfig.build` argument `area_weights` to be optional.
The config at build time will raise an error if a `type` requiring area
weights is specified but none are provided. It seems like
`WeightedMappingLoss` is used for loss in most `fme` locations, so I
only needed to make a slight adjustment there.
- Removed `area_weights` in all model building methods because we don't
use it in the loss. We might eventually want a "dynamic" loss
definition, but I'll leave that for a future adjustment if needed.


- [x] Tests updated