Fix nino3.4 calculation for e3sm (#1733)
When running training for E3SM get warning:
2025-03-21 04:42:54,539 - root - INFO - Variable sst not found in data.
Skipping Nino3.4 index computation for this variable.
2025-03-21 04:42:54,539 - root - INFO - Variable surface_temperature not
found in data. Skipping Nino3.4 index computation for this variable.
2025-03-21 04:42:54,540 - root - INFO - Variable sst not found in data.
Skipping Nino3.4 index computation for this variable.
2025-03-21 04:42:54,540 - root - INFO - Variable surface_temperature not
found in data. Skipping Nino3.4 index computation for this variable.

The corresponding variable name in E3SM is TS. This PR adds this
variable

---------

Co-authored-by: Finn Rebassoo <rebassoo@login17.chn.perlmutter.nersc.gov>