Downscaling model prep for new dataset loading (#1730)
This PR is a component of the downscaling rewrite that removes the
assumption of a static spatial locations.

In this PR, I prepare the downscaling models for usage with an upcoming
dataset rewrite that will incorporate a `PairedBatchData` class with
attributes including batch-relevant spatial information (like topography
here). The goal was to reduce the surface area of the code handling the
topography compared to what I removed in #1709. Some things re-appear,
but I was able to limit most handling to the `models.py` and
`datasets.py`, while I highlight what will disappear with the final
updates.

Since fine topography has to be concatenated to a field with the same
shape when used, I had to introduce the idea of a registered module
expressing whether it requires "interpolated" input. That allows for the
top-level model class to potentially interpolate input tensors (and
concatenate with topography if expected) before passing it to the
underlying module. Previously, this is why all of the topography
information was propagated all the way down the prediction chain and
handled in the low-level model modules.

Changes:
- Add a placeholder `PairedBatchData` type that is expected for
training/generation
- `DataLoaderConfig` has a property expressing whether
`random_subsetting_enabled` for consistency checks
- `DonwscalingModelConfig` and `DiffusionModelConfig` now:
- has `use_fine_topography` flag designated inclusion of fine topography
in the model inputs
- fails if the model should use `fine_topography` but the underlying
module does not accept inputs interpolated to the target resolution
- `Model` and `DiffusionModel` now:
  - take in `PairedBatchData` for training and generation
- will interpolate inputs (and concatenate fine topography) if requested
- raise a `ValueError` if use of fine topography is requested but
missing from an input batch
- `ModuleRegistrySelector` and `DiffusionModuleRegistrySelector` now
include argument `expects_interpolated_input` which can override the
default defined in the global mapping `EXPECTS_INTERPOLATED`
- Removed implicit interpolation from forward call of
`UNetDiffusionModule`

- [x] Tests added

---------

Co-authored-by: Anna Kwa <annak@allenai.org>