Allow `GriddedOperations` to be passed to `Loss` (#1735)
Small change to `Stepper` and `WeightedMappingLossConfig` to allow the
passing of `GriddedOperations` to the loss builder. This is in
preparation to construct a SHT-based loss function to potentially
improve spectral biases (see https://arxiv.org/abs/2501.19374). This
also simplifies `Stepper` as it removes `area_weighted_mean` which is
already contained in `GriddedOperations`
Changes:
- `WeightedMappingLossConfig` takes in `GriddedOperations` instead of
`area_weighted_mean` for the builder

- Removes reference to `area_weighted_mean` in `Stepper`

- `test_loss.py` is updated to reflect changes

---------

Co-authored-by: Jeremy McGibbon <mcgibbon@uw.edu>