Use text file for latest beaker image (#1748)
When our dependencies change, all baseline runs need to be updated to
refer to the image with new dependencies, otherwise the run will fail
with whatever exception is caused by the image used not having that
required dependency. This PR creates a text file with this image so that
all runs can be updated without editing the files for all runs.

Changes:
- Created `latest_deps_only_image.txt` which refers to the beaker image
which has dependencies needed to run the latest main code.

- [ ] Tests added