Validate vertical coarsening indices in data processing (#1753)
This PR adds safety to our data processing scripts to ensure that we are
intentional when providing vertical coarsening indices either for the
atmosphere or land that are inconsistent with the size of the relevant
vertical dimension.

Changes:
- Modifies `compute_dataset.py` and `compute_dataset_e3smv2.py` to by
default raise an error if the atmosphere or land vertical coarsening
indices do not exactly span the relevant vertical dimension with no
overlap.
- Adds `validate_vertical_coarsening_indices` and
`validate_vertical_coarsening_indices_land` flags to
`DatasetComputationConfig`, which by default are set to `True`. These
can be used to disable this check for either component.

Sample error messages look like:
```
ValueError: Provided atmosphere vertical coarsening indices [(0, 7), (7, 10), (10, 13), (13, 16), (16, 18), (18, 22), (22, 25), (25, 32)] do not exactly span all 33 vertical levels of the input dataset and/or consist of overlapping ranges of levels. If your intent is to use incomplete or overlapping levels, you may disable this check by setting validate_vertical_coarsening_indices in DatasetComputationConfig to False. If this is not your intent, double check that you are using the proper coarsening indices and, if relevant, the vertical coordinate reference file for this reference model component configuration.
```
for the atmosphere, and
```
ValueError: Provided land vertical coarsening indices [(0, 2), (2, 6), (6, 10), (10, 15)] do not exactly span all 20 vertical levels of the input dataset and/or consist of overlapping ranges of levels. If your intent is to use incomplete or overlapping levels, you may disable this check by setting validate_vertical_coarsening_indices_land in DatasetComputationConfig to False. If this is not your intent, double check that you are using the proper coarsening indices and, if relevant, the vertical coordinate reference file for this reference model component configuration.
```
for the land.

- [x] Tests added