include next_step_input_names in forcing data (#1770)
This PR updates StepperConfig to include names required in
`next_step_input_data` argument to it's `.step` in the list of "forcing
data" names required.

For input names, if the input value is an output, it does not need to be
present after the first timestep because the model's prediction can be
used autoregressively. An exception to this happens if we need to
provide the next-step value as an additional input to force the model
output.

Changes:
- Removed unused
`SingleModuleStepperConfig.get_forcing_window_data_requirements`
- Moved `next_step_input_names` from `StepABC` to `StepConfigABC`
- Include `next_step_input_names` in the names property of
`StepperConfig.get_forcing_window_data_requirements`

- [x] Tests added

Resolves #1769