Composite downscaling evaluation (#1750)
This PR adds a wrapper around the downscaling model loaded during
evaluation so that it can be used to predict smaller patches of a
prediction region larger than the size it was trained on. An optional
evaluation configuration for a `MultipatchConfig` can be specified to
have the model outputs stitched back together into a larger region.

If the loaded data region does not neatly fit an integer number of
patches with the specified overlap, the last patch along the lat/lon
dimensions is trimmed from the start of its indices such that the
overlap with the previous slice is consistent with the rest of the
region.

Note that the compositing (if specified) is done on the model outputs
(not during diffusion steps) so the composited outputs are more for
visual convenience.

Example of a CONUS prediction composited using a model with 64x64 fine
prediction shape: https://wandb.ai/ai2cm/annak-downscaling/runs/5ysjl8ya

---------

Co-authored-by: Andre Perkins <frodre@gmail.com>