Use 4 data workers and prefetch_factor=4 in example configs (#1778)
The `num_data_workers` option on the data loader config is the # of
processes per GPU. So our current examples which use 32 are launching a
huge number of processes (256 on 8-GPU node!). Decreasing the number
does not hurt performanceand actually decreases the time for initialization
of the data loader (~80s to 13s). This speedup in init will likely
be even larger when we use `forkserver` multi-processing context as
necessary for zarr async.