Handle NaNs in target data when computing loss (#1773)
Ensure that targets don't have NaNs when computing the loss and that the
filled value in prediction and target are the same so that target NaN
regions contribute 0 to the loss.

Changes:
*
[`fme/core/loss.py`](diffhunk://#diff-ff4c6872eaf21aa3e7cc5fe360e89ba28ebc0be717ca859c6ca8623370c65c2fR67-R70):
Added logic to mask NaN values in the target tensors and replace them
with zeros in both the prediction and target tensors before computing
the loss.
*
[`fme/core/test_loss.py`](diffhunk://#diff-08de741eaaf5fa2a54d16dd45dd56d4c8076ee9fef58366af15711745b2e7c61R288-R322):
Added a new test `test_WeightedMappingLoss_with_target_nans` to verify
that the `WeightedMappingLoss` correctly handles NaN values in the
target tensors.Short description of why the PR is needed and how it
satisfies those requirements, in sentence form.

- [x] Tests added