Remove unnecessary sht_fix (#1808)
`sht_fix.py` was introduced to enable multiple forward steps during
training, but does not appear to be necessary as long as SHT objects are
put on the correct device. It was also causing issues due to its
in-place mutation of the `nn.Module` weights. This PR removes it.

Changes:
- `torch_harmonics.RealSHT` and `torch_harmonics.InverseRealSHT` are no
longer overridden at import time

- [ ] Tests added