Revert `UNMASKED_OCEAN_DERIVED_NAMES` (#1812)
Reverts the addition of `UNMASKED_OCEAN_DERIVED_NAMES` introduced by
cdd56d92 which was used to leave `ocean_heat_content` unmasked. So long
as the surface mask invalid points are a subset of deeper layer invalid
points (as is currently the case and likely will be for the ocean), this
is unnecessary.