Use fme.ace for baseline configs test imports (#1814)
Use `fme.ace` namespace for testing baseline configs. This will make it
easier to distinguish downscaling train/evaluator configs in this test
module when we add these baseline configs. Also it is good practice to
not import from deeply nested modules in this test module, since it is
external to `fme` package so should just use "public" API (admittedly
"public" is not that well defined for `fme`).

Changes:
- use `fme.ace.TrainConfig` instead of importing `TrainConfig` from
`fme.ace.train.train_config`, and similarly for inference/evaluator
configs