Process 3d radiative heating rates in abrupt 4xCO2 ensemble (#1810)
We output the 3d radiative heating rates in the abrupt 4xCO2 ensemble we
ran with SHiELD. These will be useful in assessing the accuracy of ACE
in configurations in which it predicts these fields, so we update our
data processing configuration to include them. This also effectively
converts the dataset to zarr v3. Since this is a drop-in replacement for
the previous dataset with more fields, we simply replace the data at the
previous paths.

Changes:
- Updates our data processing configuration for the abrupt 4xCO2
ensemble to include
`tendency_of_air_temperature_due_to_longwave_heating` and
`tendency_of_air_temperature_due_to_shortwave_heating` as additional 3d
variables.

Argo job: `compute-fme-dataset-ensemble-qk6kw`.