Add gantry script for downscaling baseline (#1683)
Changes:
- Adds a baseline training configuration and gantry run script for
downscaling. The run is configured for using the Augusta cluster and GCS
datasets since it's easier to find available GPUs there.
- Updated the configuration tests for separate handling of the
downscaling files
- Updated `CheckpointConfig` to lazily load the checkpoint dictionary
(so validation configuration checks don't fail)


The training is configured to divide the CONUS into patches with a
random offset. The evaluation scripts evaluate over previous regions (FL
and CA) for comparison to previous results, and the entire CONUS (w/o
patching).

---------

Co-authored-by: Andre Perkins <frodre@gmail.com>