Switch Optional[...] to ... | None in fme.core (#1839)
Since python 3.10 we can use the bar `|` operator in place of `Union`.
This is also recommended by Google python style docs. As a first pass
this PR just changes instances of `Optional[...]` to `... | None` in
`fme.core` only. We can apply this change to other modules, and make
the `Union` changes in future PRs.