Don't use a 0-based epoch counter in wandb logs (#1849)
Currently when we log the "epoch" number to wandb, it is logged as 0
after the first epoch, 1 after the second, etc. This doesn't make sense,
since this number should represent the # of epochs of training
completed. Furthermore, this behavior makes it more difficult to add a
validation step before any training (see #359) which *actually* should
log with epoch=0.

Changes:
- use a single `self._epoch` counter within generic `Trainer` which gets
incremented at start of epoch loop
- the epoch logged to wandb will be 1-based instead of 0-based as it is
currently
- the "end of epoch callback" will also get this 1-based epoch instead
of the current 0-based epoch (note this argument is unused in the only
existing end-of-epoch callback that is implemented)
- the `set_epoch` method of the sampler will also now get the 1-based
epoch instead of 0-based (this just sets the seed for shuffling in each
epoch, so just need it to be different at each epoch)

- [x] Tests added