Add dataset output to downscaling evaluation (#1858)
Previously, there were some aggregator outputs saved to file that were
removed in recent reworks of the aggregator. This PR adds back saving of
select metrics that are more difficult to generate to be used for
plotting.

Changes:
- Reduces the batch size on the augusta baseline eval script due to OOM
- Add `get_dataset` methods to `GenerationAggregator`, `Aggregator`,
`ComparedDynamicHistogramsAdapter`, `ZonalPowerSpectrumAggregator`, and
`MeanMapAggregator` to facilitate saving a netcdf of metric outputs.

- [x] Tests added