Ensure doctest examples are actually running (#1860)
Currently doctest is not actually running on some of our examples
because they are invalid (see #1525 ).

In this PR we add a custom `IGNORE_OUTPUT` flag to `conf.py` so that
these examples run as expected. They still do not require any output to
be included in the example (as desired) but they will fail if the class
definitions change such that the examples are invalid.

I did this based on a suggestion in
https://stackoverflow.com/questions/3862274/doctests-how-to-suppress-ignore-output,
but we could also add `doctest_plus` dependency as suggested in #1525 .

Also made another minor change to `DataLoaderConfig` to remove some
warnings.

Resolves #1525