Add pyupgrade to ruff linter rules (#1869)
This will help us stick to modern python functionality. Add "pyupgrade"
ruleset to ruff linter: https://docs.astral.sh/ruff/rules/#pyupgrade-up

Changes:
- add `UP` to `tool.ruff.lint.select`
- ignore `UP` for files in `scripts/*`
- fix corresponding issues

Note:
We have some classes which have an attr named `type` which can conflict
with the builtin `type` that is now supposed to be used instead of
`Type`. So I put in some `ignore` statements to allow us to still use
this deprecated `Type` for now. This could be resolved other ways, but
this seemed the simplest.