Variable metadata from stepper in inference (#1873)
Allows the inference and evaluator routines to access variable metadata
that was used by the stepper in training. This is useful particularly
for inference where the forcing dataset doesn't contain output
variables, and so previously their metadata was missing.

Changes:
- Stepper serialization now includes training dataset variable metadata
in `DatasetInfo`. This metadata is used preferentially over inference
dataset metadata. If there are conflicts in metadata between stepper and
dataset a warning is logged.