Use instances for config in test_train (#1884)
This PR updates the yaml generation in `fme/ace/test_train.py` to make
use of explicit instances for configuration classes which then get
converted into yaml, rather than directly using string surgery to
construct a yaml file. This will make it easier to refactor this
configuration, and make it clearer when the test needs to be updated (as
mypy can check it).

Changes:
- `XarrayDataConfig.torch_dtype` is now a property method instead of a
dataclass attribute, to avoid issues with serializing its type

- [ ] Tests added