Fix bug in how `DepthCoordinate` is initialized (#1883)
With the introduction of `MaskProvider`
(https://github.com/ai2cm/full-model/pull/1818), variable-specific masks
like `"mask_sst"` can be added to the data. This PR fixes an error that
can occur when a `DepthCoordinate` is initialized for a dataset that
includes these new types of masks.

Changes:
- `fme.core.dataset.xarray._get_vertical_coordinate` now checks for the
regex `"mask_(\d+)$"` rather than the simple `"mask_"` substring search
we had before.