Inference default variable metadata (#1880)
As a follow-on to https://github.com/ai2cm/full-model/pull/1873 adds
default variable metadata for use in inference and evaluation,
particularly for when the serialized stepper lacks metadata
(eg for our published checkpoints). The metadata comes from
the dataset used for training ACE2-ERA5.

Changes:
- Default variable metadata is used for output variables, if it is not
found in the stepper or the forcing dataset. A warning is logged when
default metadata is in use about the potential for mismatches with
the stepper's training data.

- [X] Tests added