Add test of conversion code in Stepper.from_state (#1905)
Code for converting old checkpoints (including all checkpoints we've
published) has no coverage. It used to have coverage from the evaluator
backwards compatibility test, but we have been pretty liberal about
updating the checkpoint the test so that test no longer exercises the code.
This PR adds a specific unit test using a config that is similar to what
our published checkpoints (ACE-climSST, ACE2) use.

This came up because I noticed while working on #1903 that the relevant
code was untested.