Get `Stepper` output post-processing func from `DatasetInfo.mask_provider` (#1888)
Formerly, the `post_process_func` passed to `Stepper` init was built
with a method on the `dataset_info.vertical_coordinate` attribute. In
this PR, `post_process_func` is instead built with a method on the
`dataset_info.mask_provider` attribute. As a consequence, any dataset
with appropriate variables with the prefix `"mask_"` can now take
advantage of this automatic support for masking outputs, whereas before
only datasets that support a `DepthCoordinate` could use this feature.

Changes:
- Adds the `build_output_masker` method to `MaskProvider`, which returns
an instance of `StaticMasking`.
- Removed `VerticalCoordinate.build_post_process_function` and renamed
`DepthCoordinate.build_post_process_function` to `build_output_masker`.