Fix test of patch shuffling (#1957)
Previously only one element of the shuffled patches was compared, so
there was a higher chance it could match between shuffled batches. This
fixes the test to compare the full shuffled order. I'm not sure if this
was behind the test failure
[here](https://github.com/ai2cm/full-model/actions/runs/15545856907/job/43767059793?pr=1949)
as both CPU and GPU tests somehow failed the shuffle=True case, maybe it
was just very unlucky and they both hit the same bug.