Option to turn off inline inference during training (#1949)
For convenience, we sometimes don't want to do inline inference during
training. For example, when benchmarking multi-node training, we're not
concerned about which checkpoint is best and thus do not need inline
inference. While it is possible to specify a list of inference starting
index to match the number (or multiple) of GPUs and just set the number
of forward steps to 1, it is convenient to have an option to not do it
at all.

Changes:
- In `TrainConfig`, if no inference is specified, no inline inference is
done

- [-] Tests added

Beaker training run w/o inline inference:
[here](https://beaker.allen.ai/orgs/ai2/workspaces/ace/work/01JWYKVR6V1AG006AW3D1JETKX?taskId=01JWYKVR6ZF9MCBNKDW6GMWXN9&jobId=01JWYNA1YVHWG35MFSGZJP0VJB)

---------

Co-authored-by: Anna Kwa <annak@allenai.org>
Co-authored-by: Brian Henn <brianhenn@allenai.org>