Compatibility with dataset missing vertical coordinate (#1946)
Makes `NullVerticalCoordinate` compatible with real vertical
coordinates, such that a dataset with only 2D variables is considered
compatible with a dataset with a real vertical coordinate. For example,
this happens when we run inference with 2D forcing variables for a
stepper trained on a vertical variable dataset. Also makes the inference
integration test check this specific case.