Stepper training history (#1967)
Adds a training history attribute to the Stepper class, which is appended to in training jobs when any stepper checkpoint is saved. Currently this attribute contains the git SHA and wandb ID of the training job. Training that involves fine-tuning from a pre-trained set of model weights adds the pre-training history to the history.

Changes:
- Steppers now save and load with training history included; this will be empty upon loading for pre-existing saved steppers.