Fix downscaling output shape determination during eval  (#1984)
A recent PR adjusted the output shape determination downscaling to be
based off of the `fine_shape` attribute. However, this attribute assumes
that we have also require a fixed input coarse horizontal resolution,
which is not a requirement for our downscaling models. This caused a few
of my recent evals to fail where I was trying to output the entire CONUS

Changes:
- removed `fine_shape` in place of `get_fine_shape()` to indicate it's a
dynamic measure
- moved the normalizing into the `_generate` helper function (also
renamed) to keep the code closer in proximity and now pass three outputs
`generated`, `generated_norm`, and `latent_steps`
- Added a test for arbitrary input size
- Renamed variables something other than "samples" since that's
routinely overloaded

- [x] Tests added