Make default parameter_init behavior be from config (#1975)
Makes `StepperConfig.get_stepper` build the parameter initializer from
config instead of requiring that it be passed as an argument. In cases
where no parameter initialization is desired (e.g., on inference), the
option of passing `apply_paramater_init=False` is added. This avoids the
need to have a separate build call of the parameter initializer in
training routines, which was introduced in #1945 and was bug-prone.

Resolves https://github.com/ai2cm/full-model/issues/1973

---------

Co-authored-by: James Duncan <jpduncan@berkeley.edu>