DatasetInfo uses HorizontalCoordinates and not gridded ops (#1903)
This PR changes the `DatasetInfo` class to now take the
`HorizontalCoordinates` on init instead of the `img_shape` and
`GriddedOperations`. This is necessary to properly describe the training
dataset and enable validation of coordinates at inference time. For
backwards compatibility, `img_shape` and `GriddedOperations` can be
provided, but not at the same time as the `HorizontalCoordinates`.

Changes:
- `DatasetInfo` class can now take the `HorizontalCoordinates` on init
instead of the `img_shape` and `GriddedOperations`. When training new
models, the `DatasetInfo` will be generated with the horizontal
coordinate.
- `HorizontalCoordinates` now has a `.shape` attr which can be used in
place of the old `DatasetInfo.img_shape` attr
- `HorizontalCoordinates` how has a `to_state` method and can be
serialized with `SerializableHorizontalCoordinate`

- [x] Tests added

Resolves #1433 and mostly addresses #1879